test: split oversized route and workout tests
